requirements.txt
asgiref==3.2.7
coverage==5.1
Django==3.0.5
django-crispy-forms==1.9.0
pytz==2019.3
selenium==3.141.0
sqlparse==0.3.1
urllib3==1.25.9
